fix(DifftestBundle): fix incorrect C++ type declarations during code generation
When generating `arrayType`, not all 8-bit-aligned types are valid in C++.

For example, if someone defines a 41-bit-wide signal, it may result in
generating an unsupported C++ type like uint48_t instead of a valid uint64_t.

This change ensures that for any given bit width, a C++-supported type is generated.